+++ lib/persistence/entities/parsed_pantry_item.dart
@@ -7,8 +7,8 @@ class ParsedPantryItem {
 
   ParsedPantryItem({this.food});
 
-  int getTotalRemainingQuantity() {
-    int total = 0;
+  double getTotalRemainingQuantity() {
+    double total = 0;
     for (PantryItem item in this.items) {
       total += item.leftQuantity;
     }
@@ -18,8 +18,6 @@ class ParsedPantryItem {
   int getStockQuantity() {
     return this.items.length;
   }
-
-
 
   factory ParsedPantryItem.fromPantryItemList(List<PantryItem> items) {
     ParsedPantryItem parsed = ParsedPantryItem();
